Move the key management outside the enclave package, as it is not strictly only to be used by the Enclave
