+++ nexus-app/src/main/java/com/github/nexus/keys/KeyManager.java
@@ -1,4 +1,4 @@
-package com.github.nexus.enclave.keys;
+package com.github.nexus.keys;
 
 import com.github.nexus.nacl.Key;
 import com.github.nexus.nacl.KeyPair;
